remove cuisine and category

+++ cookbook/src/main/java/cookbook/controllers/CategoryAndCuisineController.java
@@ -20,6 +20,7 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import cookbook.database.CategoryRepository;
 import cookbook.database.CuisineRepository;
+import cookbook.database.RecipeRepository;
 import cookbook.exception.ResourceNotFoundException;
 import cookbook.models.Category;
 import cookbook.models.Cuisine;
@@ -41,29 +42,31 @@ public class CategoryAndCuisineController {
 
 	@Autowired
 	private CuisineRepository cuisineRepository;
-	
+
 	@Autowired
 	private RecipeService recipeService;
 
+	@Autowired
+	private RecipeRepository recipeRepository;
+
 	@GetMapping("/category/{category}")
-	public PagedResponse<RecipeResponse> getAllRecipesByCategory(
-			@CurrentUser UserPrincipal currentUser,
+	public PagedResponse<RecipeResponse> getAllRecipesByCategory(@CurrentUser UserPrincipal currentUser,
 			@RequestParam(value = "page", defaultValue = AppConstants.DEFAULT_PAGE_NUMBER) int page,
 			@RequestParam(value = "size", defaultValue = AppConstants.DEFAULT_PAGE_SIZE) int size,
-			@PathVariable(value = "category") String category){
+			@PathVariable(value = "category") String category) {
 		return recipeService.getByCategory(currentUser, page, size, category);
-		
+
 	}
-	
+
 	@GetMapping("/category/all")
 	public List<String> getCategories() {
 		return recipeService.getAllCategories();
 	}
-	
+
 	@PostMapping("/category")
 	@PreAuthorize("hasRole('USER')")
 	public ResponseEntity<?> addCategory(@RequestParam Map<String, String> category) {
-		
+
 		if (categoryRepository.existsById(category.get("category"))) {
 			return new ResponseEntity<>(new ApiResponse(false, "Category exists!"), HttpStatus.BAD_REQUEST);
 		}
@@ -83,28 +86,30 @@ public class CategoryAndCuisineController {
 	@PreAuthorize("hasRole('USER')")
 	public ResponseEntity<?> removeCategory(@PathVariable(value = "category") String category) {
 
+		if (recipeRepository.countByCategoryName(category)<= 0) {
 		categoryRepository.findById(category)
 				.orElseThrow(() -> new ResourceNotFoundException("Such category doesn't exist", "category", category));
 
 		categoryRepository.deleteById(category);
 
 		return new ResponseEntity<>(new ApiResponse(true, "Category removed successfully"), HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(new ApiResponse(false, "Recipe with this category exists"), HttpStatus.BAD_REQUEST);
+		}
 	}
 
 	@GetMapping("/cuisine/{cuisine}")
-	public PagedResponse<RecipeResponse> getAllRecipesByCousine(
-			@CurrentUser UserPrincipal currentUser,
+	public PagedResponse<RecipeResponse> getAllRecipesByCousine(@CurrentUser UserPrincipal currentUser,
 			@RequestParam(value = "page", defaultValue = AppConstants.DEFAULT_PAGE_NUMBER) int page,
 			@RequestParam(value = "size", defaultValue = AppConstants.DEFAULT_PAGE_SIZE) int size,
-			@PathVariable(value = "cuisine") String cuisine){
+			@PathVariable(value = "cuisine") String cuisine) {
 		return recipeService.getByCousine(currentUser, page, size, cuisine);
 	}
-	
-	
+
 	@PostMapping("/cuisine")
 	@PreAuthorize("hasRole('USER')")
 	public ResponseEntity<?> addCousine(@RequestParam Map<String, String> cuisine) {
-		
+
 		if (cuisineRepository.existsById(cuisine.get("cuisine"))) {
 			return new ResponseEntity<>(new ApiResponse(false, "Cuisine exists!"), HttpStatus.BAD_REQUEST);
 		}
@@ -124,14 +129,19 @@ public class CategoryAndCuisineController {
 	@PreAuthorize("hasRole('USER')")
 	public ResponseEntity<?> removeCousine(@PathVariable(value = "cuisine") String cuisine) {
 
-		cuisineRepository.findById(cuisine)
-				.orElseThrow(() -> new ResourceNotFoundException("Such cuisine doesn't exist", "cuisine", cuisine));
+		if (recipeRepository.countByCuisineName(cuisine) <= 0) {
+			cuisineRepository.findById(cuisine)
+					.orElseThrow(() -> new ResourceNotFoundException("Such cuisine doesn't exist", "cuisine", cuisine));
 
-		cuisineRepository.deleteById(cuisine);
+			cuisineRepository.deleteById(cuisine);
 
-		return new ResponseEntity<>(new ApiResponse(true, "Cuisine removed successfully"), HttpStatus.OK);
+			return new ResponseEntity<>(new ApiResponse(true, "Cuisine removed successfully"), HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(new ApiResponse(false, "Recipe with this cuisine exists"), HttpStatus.BAD_REQUEST);
+		}
+
 	}
-	
+
 	@GetMapping("/cuisine/all")
 	public List<String> getCuisine() {
 		return recipeService.getAllCuisine();
